Split transactions table internals
